refactor(attachments): add resource-attachments
