+++ src/attachments/attachments.module.ts
@@ -1,19 +1,14 @@
 import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
 import { PermissionsModule } from 'omniboxd/permissions/permissions.module';
 import { AttachmentsController } from 'omniboxd/attachments/attachments.controller';
 import { AttachmentsService } from 'omniboxd/attachments/attachments.service';
 import { MinioModule } from 'omniboxd/minio/minio.module';
-import { ResourceAttachment } from './entities/resource-attachment.entity';
+import { ResourceAttachmentsModule } from 'omniboxd/resource-attachments/resource-attachments.module';
 
 @Module({
   exports: [AttachmentsService],
   providers: [AttachmentsService],
   controllers: [AttachmentsController],
-  imports: [
-    TypeOrmModule.forFeature([ResourceAttachment]),
-    PermissionsModule,
-    MinioModule,
-  ],
+  imports: [PermissionsModule, MinioModule, ResourceAttachmentsModule],
 })
 export class AttachmentsModule {}
